refactor(page-markdown-endpoint): consolidate into features/page-markdown and register declaratively
Move the route handlers, resolution/builder services, and the shared
.md URL / alternate-link utils under features/page-markdown/ (client-
safe barrel at the feature root, server surface behind ./server),
following the feature-based layout and the growi-vault / rate-limiter
precedent in routes/index.js. The responder was already Express-free,
so it now lives as a service; the pure builders leave src/server/routes
and its route-top-level-guard constraint.

Registration switches from an opaque app.use(routerFactory()) to the
route-table style: app.get('/*', gate, accessTokenParser, loginRequired,
respond) right above the catch-alls, reusing the middleware instances
routes/index.js already composes -- the GET verb, pattern, and authz
chain are now visible in the route table. Behavior is unchanged (pinned
by the route-level integ suite, which mirrors the same composition).
