+++ apps/app/src/features/page-markdown/server/index.ts
@@ -1,0 +1,4 @@
+export {
+  createPageMarkdownHandlers,
+  type PageMarkdownHandlers,
+} from './routes/page-markdown';
