+++ apps/app/src/features/page-markdown/server/services/build-page-markdown.ts
@@ -48,8 +48,6 @@ export interface PageMarkdownInput {
 
 const EMPTY_PAGE_NOTICE = 'This page has no content yet.';
 
-// Template literal (not .join) so the initializer stays on the
-// route-top-level-guard allowlist (no top-level call expressions).
 const ERROR_GUIDANCE = `If you believe this page exists and you have access to it, retry with an
 authenticated request (a logged-in session or a Personal Access Token),
 or use the GROWI MCP server, which can fetch this page on your behalf
